// src/models/regions/locations/spawn/traits/settlements/factions/types.ts
export type settlement__faction =
  | 'artisan guilds'
  | 'criminal bosses'
  | 'colonial outpost'
  | 'colonial embassies'
  | 'druidic circle'
  | 'foreign enclave'
  | 'knightly order'
  | 'legendary figure'
  | 'major temple'
  | 'mercenary company'
  | 'merchants guild'
  | 'missionary activity'
  | 'nomadic populace'
  | 'pirate enclave'
  | 'prestigious academy'
  | 'prominent monastery'
  | 'secret police'
  | 'sorcerous cabal'
  | 'templar order'
  | 'volunteer militia'